adjusted migration sponsorship and inserted sponsorship seeder

+++ BoolBnb/database/migrations/2021_12_02_135903_create_sponsorships_table.php
@@ -16,7 +16,7 @@ class CreateSponsorshipsTable extends Migration
         Schema::create('sponsorships', function (Blueprint $table) {
             $table->id();
             $table->decimal('price', 7, 2);
-            $table->tinyInteger('hour_duration');
+            $table->smallInteger('hour_duration');
             $table->timestamps();
         });
     }
